requirements.txt
asgiref==3.4.1
attrs==21.2.0
Babel==2.9.1
Django==3.2.9
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.12.4
iniconfig==1.1.1
packaging==21.2
pluggy==1.0.0
psycopg==3.0.4
psycopg2-binary==2.9.2
py==1.11.0
py-moneyed==2.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.0
